+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * added: coverage-plugin helper
 * added: nuclei plugin
 * added: server: ux, fuller storage datatables toolbar multi-action toolboxes
+* added: planner, 'load_standalone' stages
 * fixed: server: fix filter POST request
 * fixed: timeout for OIDC calls
 * security: bump jquery